fix env-tests
