+++ packages/core/compute/src/compute-graph.test.ts
@@ -7,7 +7,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'vitest';
 
 import { Trigger } from '@dxos/async';
 import { create, fullyQualifiedId } from '@dxos/client/echo';
-import { FunctionType } from '@dxos/functions';
+import { FunctionType } from '@dxos/functions/types';
 
 import { DetailedCellError } from '#hyperformula';
 import { TestBuilder } from './testing';
