+++ packages/core/functions/package.json
@@ -78,7 +78,7 @@
     "@dxos/schema": "workspace:*",
     "@dxos/util": "workspace:*",
     "@preact/signals-core": "^1.6.0",
-    "cron": "^3.1.6",
+    "cron-schedule": "^5.0.4",
     "effect": "3.12.12",
     "esbuild-wasm": "^0.16.14",
     "express": "^4.19.2",
