+++ packages/core/functions/src/edge/functions.ts
@@ -5,17 +5,11 @@
 import { type DID } from 'iso-did/types';
 
 import { type Client } from '@dxos/client';
-import { type ObjectMeta } from '@dxos/echo-schema';
 import { EdgeHttpClient, type EdgeIdentity } from '@dxos/edge-client';
 import { invariant } from '@dxos/invariant';
 import type { PublicKey, SpaceId } from '@dxos/keys';
 import { log } from '@dxos/log';
 import { type UploadFunctionResponseBody } from '@dxos/protocols';
-
-// TODO: use URL scheme for source?
-const FUNCTIONS_META_KEY = 'dxos.org/service/function';
-
-export const FUNCTIONS_PRESET_META_KEY = 'dxos.org/service/function-preset';
 
 export type UploadWorkerArgs = {
   client: Client;
@@ -52,37 +46,6 @@ export const uploadWorkerFunction = async ({
   return response;
 };
 
-export const getUserFunctionUrlInMetadata = (meta: ObjectMeta) => {
-  return meta.keys.find((key) => key.source === FUNCTIONS_META_KEY)?.id;
-};
-
-export const setUserFunctionUrlInMetadata = (meta: ObjectMeta, functionUrl: string) => {
-  const key = meta.keys.find((key) => key.source === FUNCTIONS_META_KEY);
-  if (key) {
-    if (key.id !== functionUrl) {
-      throw new Error('Metadata mismatch');
-    }
-  } else {
-    meta.keys.push({ source: FUNCTIONS_META_KEY, id: functionUrl });
-  }
-};
-
-export const getInvocationUrl = (functionUrl: string, edgeUrl: string, options: InvocationOptions = {}) => {
-  const baseUrl = new URL('functions/', edgeUrl);
-
-  // Leading slashes cause the URL to be treated as an absolute path.
-  const relativeUrl = functionUrl.replace(/^\//, '');
-  const url = new URL(`./${relativeUrl}`, baseUrl.toString());
-  options.spaceId && url.searchParams.set('spaceId', options.spaceId);
-  options.subjectId && url.searchParams.set('subjectId', options.subjectId);
-  url.protocol = isSecure(url.protocol) ? 'https' : 'http';
-  return url.toString();
-};
-
-const isSecure = (protocol: string) => {
-  return protocol === 'https:' || protocol === 'wss:';
-};
-
 const createEdgeIdentity = (client: Client): EdgeIdentity => {
   const identity = client.halo.identity.get();
   const device = client.halo.device;
@@ -112,9 +75,4 @@ export const incrementSemverPatch = (version: string): string => {
 
 export const publicKeyToDid = (key: PublicKey): DID => {
   return `did:key:${key.toHex()}`;
-};
-
-export type InvocationOptions = {
-  spaceId?: SpaceId;
-  subjectId?: string;
 };
