+++ packages/core/functions/src/types/index.ts
@@ -5,3 +5,4 @@
 export * from './schema';
 export * from './types';
 export * from './trace';
+export * from './url';
