+++ packages/devtools/cli/src/commands/function/list.ts
@@ -9,7 +9,7 @@ import { join } from 'node:path';
 
 import { table } from '@dxos/cli-base';
 import { Config } from '@dxos/config';
-import { type FunctionManifest } from '@dxos/functions';
+import { type FunctionManifest } from '@dxos/functions/types';
 
 import { BaseCommand } from '../../base';
